+++ src/main/java/com/l524l/vkcheckersbot/commandsmodule/CommandDeterminant.java
@@ -1,8 +1,8 @@
-package ru.mityushin.responder.commandsmodule;
+package com.l524l.vkcheckersbot.commandsmodule;
 
 import com.vk.api.sdk.objects.messages.Message;
-import ru.mityushin.responder.commandsmodule.commands.Command;
-import ru.mityushin.responder.exceptions.UnknownCommand;
+import com.l524l.vkcheckersbot.commandsmodule.commands.Command;
+import com.l524l.vkcheckersbot.exceptions.UnknownCommand;
 import java.util.Collection;
 
 public class CommandDeterminant {
