+++ src/main/java/com/l524l/vkcheckersbot/commandsmodule/CommandManager.java
@@ -1,9 +1,9 @@
-package ru.mityushin.responder.commandsmodule;
+package com.l524l.vkcheckersbot.commandsmodule;
 
+import com.l524l.vkcheckersbot.commandsmodule.commands.*;
+import com.l524l.vkcheckersbot.service.VkMessageSenderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import ru.mityushin.responder.commandsmodule.commands.*;
-import ru.mityushin.responder.service.VkMessageSenderService;
 
 import java.util.HashSet;
