+++ src/main/java/com/l524l/vkcheckersbot/commandsmodule/commands/Command.java
@@ -1,7 +1,7 @@
-package ru.mityushin.responder.commandsmodule.commands;
+package com.l524l.vkcheckersbot.commandsmodule.commands;
 
+import com.l524l.vkcheckersbot.service.VkMessageSenderService;
 import com.vk.api.sdk.objects.messages.Message;
-import ru.mityushin.responder.service.VkMessageSenderService;
 
 public abstract class Command {
     protected final VkMessageSenderService messageSenderService;
