+++ src/main/java/com/l524l/vkcheckersbot/commandsmodule/commands/Remove.java
@@ -1,11 +1,11 @@
-package ru.mityushin.responder.commandsmodule.commands;
+package com.l524l.vkcheckersbot.commandsmodule.commands;
 
+import com.l524l.vkcheckersbot.checkers.NewCheckersBoard;
 import com.vk.api.sdk.exceptions.ApiException;
 import com.vk.api.sdk.exceptions.ClientException;
 import com.vk.api.sdk.objects.messages.Message;
-import ru.mityushin.responder.checkers.NewCheckersBoard;
-import ru.mityushin.responder.exceptions.BadParameterDetected;
-import ru.mityushin.responder.service.VkMessageSenderService;
+import com.l524l.vkcheckersbot.exceptions.BadParameterDetected;
+import com.l524l.vkcheckersbot.service.VkMessageSenderService;
 
 public class Remove extends Command {
     public Remove(VkMessageSenderService messageSenderService, String name) {
