+++ src/main/java/com/l524l/vkcheckersbot/controller/CallbackController.java
@@ -1,7 +1,9 @@
-package ru.mityushin.responder.controller;
+package com.l524l.vkcheckersbot.controller;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
+import com.l524l.vkcheckersbot.config.VKGroupActor;
+import com.l524l.vkcheckersbot.service.CallbackApiHandler;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,8 +12,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-import ru.mityushin.responder.config.VKGroupActor;
-import ru.mityushin.responder.service.CallbackApiHandler;
 
 @Controller
 @RequestMapping(value = "/callbacks")
