+++ src/main/java/com/l524l/vkcheckersbot/exceptions/UnknownCommand.java
@@ -1,4 +1,4 @@
-package ru.mityushin.responder.exceptions;
+package com.l524l.vkcheckersbot.exceptions;
 
 public class UnknownCommand extends Exception {
     public UnknownCommand(){
