+++ src/main/java/com/l524l/vkcheckersbot/service/CallbackApiHandler.java
@@ -1,14 +1,14 @@
-package ru.mityushin.responder.service;
+package com.l524l.vkcheckersbot.service;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
+import com.l524l.vkcheckersbot.config.VKGroupActor;
 import com.vk.api.sdk.callback.CallbackApi;
 import com.vk.api.sdk.objects.callback.ConfirmationMessage;
 import com.vk.api.sdk.objects.messages.Message;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ru.mityushin.responder.commandsmodule.Commander;
-import ru.mityushin.responder.config.VKGroupActor;
+import com.l524l.vkcheckersbot.commandsmodule.Commander;
 
 @Service
 public class CallbackApiHandler extends CallbackApi {
